+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/internal/DUnitHost.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.test.dunit.standalone;
+package org.apache.geode.test.dunit.internal;
 
 import java.io.IOException;
 import java.rmi.AccessException;
